//server is now able to run with inputs from config file
-update: clientBodySize use size_t as new type due to insufficient space
-exception handler move to main.cpp including for parsing exceptions
-include back most part of cfgRoute due to accident removal when resolving conflict issues
-remove or comment away debug used functions
-actual main() done
-example config file provided

todo: parsing - inline content...

+++ includes/Http.hpp
@@ -18,7 +18,7 @@ class Http
         vector<cfgRoute> routes;
         bool autoindex;
         string indexFile;
-        int bodySize;
+        size_t  bodySize;
 		map<int,string>		errorCodeMap;
         string redirectPath;
         string uplaodPath;
